Move precompiles to its own module
